refactor: extract TooltipPositionProps into types.ts
Extract the TooltipPositionProps interface from <script module> blocks
into dedicated types.ts files in three component directories:
- app/frontend/src/lib/components/app/tooltips/
- plugins/idah-image/frontend/src/lib/components/ui/Tooltips/
- plugins/idah-video/frontend/src/lib/components/ui/Tooltips/

This matches the established convention of colocating types with their
related components. No behavior change.

Refs: CU-869d9wuza
