+++ app/frontend/src/lib/components/app/tooltips/KbdTooltipButton.svelte
@@ -9,7 +9,7 @@
 
   import type { ButtonProps } from "$lib/components/ui/button/button.svelte";
   import type { IIdahDriverV2 } from "$lib/plugin/v2/types";
-  import type { TooltipPositionProps } from "$lib/components/app/tooltips/tooltips.svelte";
+  import type { TooltipPositionProps } from "$lib/components/app/tooltips/types";
 
   // Props
   interface Props extends ButtonProps, TooltipPositionProps {
