+++ plugins/idah-image/frontend/src/lib/components/ui/Tooltips/TooltipButton.svelte
@@ -6,7 +6,7 @@
   import Button from "$lib/components/ui/Button/Button.svelte";
 
   import type { ButtonProps } from "$lib/components/ui/Button/button-variants";
-  import type { TooltipPositionProps } from "$lib/components/ui/Tooltips/Tooltips.svelte";
+  import type { TooltipPositionProps } from "$lib/components/ui/Tooltips/types";
 
   // Props
   interface Props extends ButtonProps, TooltipPositionProps {
