+++ plugins/idah-image/frontend/src/lib/components/ui/Tooltips/types.ts
@@ -1,0 +1,6 @@
+export interface TooltipPositionProps {
+  delayDuration?: number;
+  align?: "start" | "center" | "end";
+  side?: "top" | "right" | "bottom" | "left";
+  onOpenChange?: (open: boolean) => void;
+}
